Renamed `template_parser` to `template`.
