+++ lib/squib/template.rb
