+++ spec/template_spec.rb
